Reworked how stack traces are rendered
