+++ src/Exceptionless.Web/ClientApp/package.json
@@ -49,11 +49,9 @@
 		"@exceptionless/browser": "^3.0.4",
 		"@iconify-json/mdi": "^1.1.54",
 		"@tanstack/svelte-table": "^8.10.3",
-		"@types/dompurify": "^3.0.3",
 		"@web3-storage/parse-link-header": "^3.1.0",
 		"class-validator": "^0.14.0",
 		"daisyui": "^3.8.3",
-		"dompurify": "^3.0.6",
 		"oidc-client-ts": "^2.3.0",
 		"svelte-local-storage-store": "^0.6.3",
 		"svelte-time": "^0.8.0",
